requirements.txt
langchain==0.1.13
langchain-google-genai==0.0.11
google-generativeai==0.4.1
chromadb==0.4.24
pypdf==4.0.1
python-dotenv==1.0.0
pydantic==2.5.0
tqdm==4.67.0
